Remove duplication from elections user roles forms
